dao: fix tests
With the various DB drivers updates for 0.23.x, the dao code
had to be updated.
