+++ src/main/java/org/killbill/billing/plugin/analytics/reports/sql/Metadata.java
@@ -68,13 +68,9 @@ public class Metadata {
     private final Map<String, Table> tablesCache = new ConcurrentHashMap<String, Table>();
     private final Map<String, Map<String, List<Object>>> distinctValuesCache = new ConcurrentHashMap<String, Map<String, List<Object>>>();
 
-    public Metadata(final Set<String> reportsTables, final DataSource dataSource) {
-        this(reportsTables, dataSource, SQLDialect.MYSQL);
-    }
-
-    public Metadata(final Set<String> reportsTables, final DataSource dataSource, final SQLDialect sqlDialect) {
+    public Metadata(final Set<String> reportsTables, final DataSource dataSource) throws SQLException {
         this.reportsTables = reportsTables;
-        this.context = DSL.using(dataSource, sqlDialect, JooqSettings.defaults(sqlDialect));
+        this.context = JooqSettings.buildDslContext(dataSource);
         primeCaches();
     }
 
@@ -213,6 +209,11 @@ public class Metadata {
             schemaName = context.connectionResult(new ConnectionCallable<String>() {
                 @Override
                 public String run(final Connection connection) throws Exception {
+                    final String schema = connection.getSchema();
+                    if (schema != null) {
+                        return schema;
+                    }
+                    // MariaDB getSchema() returns null
                     return connection.getCatalog();
                 }
             });
