+++ src/main/resources/org/killbill/billing/plugin/analytics/ddl.sql
@@ -1,4 +1,5 @@
 /*! SET default_storage_engine=INNODB */;
+/*! SET SQL_MODE=(SELECT CONCAT(@@SQL_MODE,',ANSI_QUOTES')) */;
 
 -- Subscription events
 drop table if exists analytics_subscription_transitions;
@@ -1210,7 +1211,7 @@ create table analytics_account_fields (
   record_id serial unique
 , custom_field_record_id bigint /*! unsigned */ default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
@@ -1234,7 +1235,7 @@ create table analytics_bundle_fields (
 , bundle_id varchar(36) default null
 , bundle_external_key varchar(255) default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
@@ -1259,7 +1260,7 @@ create table analytics_invoice_fields (
 , custom_field_record_id bigint /*! unsigned */ default null
 , invoice_id varchar(36) default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
@@ -1282,7 +1283,7 @@ create table analytics_invoice_payment_fields (
 , custom_field_record_id bigint /*! unsigned */ default null
 , invoice_payment_id varchar(36) default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
@@ -1305,7 +1306,7 @@ create table analytics_payment_fields (
 , custom_field_record_id bigint /*! unsigned */ default null
 , payment_id varchar(36) default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
@@ -1328,7 +1329,7 @@ create table analytics_payment_method_fields (
 , custom_field_record_id bigint /*! unsigned */ default null
 , payment_method_id varchar(36) default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
@@ -1351,7 +1352,7 @@ create table analytics_transaction_fields (
 , custom_field_record_id bigint /*! unsigned */ default null
 , transaction_id varchar(36) default null
 , name varchar(64) default null
-, value varchar(255) default null
+, "value" varchar(255) default null
 , created_date datetime default null
 , created_by varchar(50) default null
 , created_reason_code varchar(255) default null
